requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
requests==2.32.3
soupsieve==2.3.1
urllib3==1.26.8
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.2
todoist-api-python==2.1.7